netdev-dpdk: Fix unintentional integer overflow 575138
575138 Unintentional integer overflow
The expression's value may not be what the programmer intended, because the
expression is evaluated using a narrow (i.e. few bits) integer type.

In fill_meter_profile(): An integer overflow occurs,
with the result converted to a wider integer type (CWE-190)

Coverity issue: CID 575138
Issue: OVS-3364094
Fixes: 28ad235b35 ("netdev-dpdk: Introduce netdev-dpdk metes API")

+++ lib/netdev-dpdk.c
@@ -5707,8 +5707,9 @@ fill_meter_profile(struct rte_mtr_meter_profile *profile,
     } else {
         profile->packet_mode = 0;
         /* Convert from kilobits per second to bytes per second */
-        profile->srtcm_rfc2697.cir = config->bands[0].rate * 125;
-        profile->srtcm_rfc2697.cbs = config->bands[0].burst_size * 125;
+        profile->srtcm_rfc2697.cir = ((uint64_t) config->bands[0].rate) * 125;
+        profile->srtcm_rfc2697.cbs =
+            ((uint64_t) config->bands[0].burst_size) * 125;
     }
 }
